+++ examples/colorBlind.cpp
@@ -46,10 +46,9 @@ int main(void)
 	ppm = new mb::PostProcessMaterial(buffer.str().c_str());
 
 	mb::TexOptions opts;
-	mb::Texture2D* tex = new mb::Texture2D(opts, MB_TEXTURE_ASSETS + std::string("/Dundus_Square.jpg"));
-	tex->bind(0);
+	mb::Texture* tex = new mb::Texture2D(opts, MB_TEXTURE_ASSETS + std::string("/Dundus_Square.jpg"));
 
-	ppm->addUniform("tex", new mb::Uniform(mb::Integer, 0));
+	ppm->addUniform("tex", new mb::Uniform(mb::TextureSampler, tex));
 	ppm->addUniform("mode", new mb::Uniform(mb::Integer, 0));
 
 	engine->run(renderFunc);