+++ mb/Helpers.cpp
@@ -27,6 +27,51 @@
 
 namespace mb
 {
+	PointMaterial::PointMaterial()
+		: Material()
+	{
+		_uniforms["projection"] = new Uniform(Matrix4);
+		_uniforms["view"] = new Uniform(Matrix4);
+		_uniforms["model"] = new Uniform(Matrix4);
+		_uniforms["color"] = new Uniform(Vector3, Vect3(1.0f, 1.0f, 1.0f));
+		_uniforms["size"] = new Uniform(Float, 1.0f);
+		_uniforms["sizeAttenuation"] = new Uniform(Boolean, false);
+		_uniforms["opacity"] = new Uniform(Float, 1.0f);
+
+		const char* vsShader = "#version 330\n"
+			"layout(location = 0) in vec3 position;"
+			"uniform mat4 projection;"
+			"uniform mat4 view;"
+			"uniform mat4 model;"
+			"uniform float size;"
+			"uniform bool sizeAttenuation;"
+			"void main() {"
+			"	vec4 pointPosition = view * model * vec4(position, 1.0);\n"
+			"	if (sizeAttenuation) \n"
+			"		gl_PointSize = size * (scale / length(pointPosition.xyz));\n"
+			"	else\n"
+			"		gl_PointSize = size;\n"
+			"    gl_Position = projection * pointPosition;\n"
+			"}";
+		const char* fsShader = "#version 330\n"
+			"out vec4 fragColor;\n"
+			"uniform float opacity;\n"
+			"uniform vec3 color;\n"
+			"void main() {\n"
+			"	fragColor = vec4(color, opacity);\n"
+			"}\n";
+		_program.loadFromText(vsShader, fsShader);
+		_program.compileAndLink();
+		_program.autocatching();
+
+		/*
+			fragColor = fragColor * texture( tex, vec2( gl_PointCoord.x, 1.0 - gl_PointCoord.y ) );
+			fragColor.a = opacity;
+		*/
+
+		// TODO: Enable blending
+		// blend_src_rgb = SRC_ALPHA
+	}
 	MoveComponent::MoveComponent()
 		: mb::Component()
 		, _delta(0.01f)