+++ mb/scene/Component.hpp
@@ -43,6 +43,7 @@ namespace mb
 		MB_API
 		virtual void start();
 		MB_API
+			// TODO: USE const float& dt
 		virtual void update(float dt);
 		MB_API
 		Node* getNode() const;